[AND-543] Rename MessageSender to SendMessageInterceptor.
