+++ stream-chat-android-client/src/debug/java/io/getstream/chat/android/client/di/ChatModule.kt
@@ -20,11 +20,11 @@ import android.content.Context
 import androidx.lifecycle.Lifecycle
 import io.getstream.chat.android.client.api.ChatClientConfig
 import io.getstream.chat.android.client.debugger.ChatClientDebugger
+import io.getstream.chat.android.client.interceptor.SendMessageInterceptor
 import io.getstream.chat.android.client.notifications.handler.NotificationHandler
 import io.getstream.chat.android.client.parser.ChatParser
 import io.getstream.chat.android.client.scope.ClientScope
 import io.getstream.chat.android.client.scope.UserScope
-import io.getstream.chat.android.client.sender.MessageSender
 import io.getstream.chat.android.client.token.TokenManager
 import io.getstream.chat.android.client.transformer.ApiModelTransformers
 import io.getstream.chat.android.client.uploader.FileTransformer
@@ -46,7 +46,7 @@ internal class ChatModule(
     apiModelTransformers: ApiModelTransformers,
     fileTransformer: FileTransformer,
     uploader: FileUploader?,
-    messageSender: MessageSender?,
+    sendMessageInterceptor: SendMessageInterceptor?,
     tokenManager: TokenManager,
     customOkHttpClient: OkHttpClient?,
     clientDebugger: ChatClientDebugger?,
@@ -62,7 +62,7 @@ internal class ChatModule(
     apiModelTransformers,
     fileTransformer,
     uploader,
-    messageSender,
+    sendMessageInterceptor,
     tokenManager,
     customOkHttpClient,
     clientDebugger,
