+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/api/ProxyChatApiTest.kt
@@ -16,9 +16,9 @@
 
 package io.getstream.chat.android.client.api
 
+import io.getstream.chat.android.client.interceptor.SendMessageInterceptor
 import io.getstream.chat.android.client.scope.ClientScope
 import io.getstream.chat.android.client.scope.UserScope
-import io.getstream.chat.android.client.sender.MessageSender
 import io.getstream.chat.android.client.utils.RetroSuccess
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.randomMessage
@@ -58,12 +58,12 @@ internal class ProxyChatApiTest {
         val channelId = randomString()
         val message = randomMessage()
         val originalApi = mock<ChatApi>()
-        val messageSender = mock<MessageSender>()
-        val proxyApi = ProxyChatApi(originalApi, UserScope(ClientScope()), messageSender)
+        val sendMessageInterceptor = mock<SendMessageInterceptor>()
+        val proxyApi = ProxyChatApi(originalApi, UserScope(ClientScope()), sendMessageInterceptor)
         // When
         proxyApi.sendMessage(channelType, channelId, message).await()
         // Then
         verify(originalApi, never()).sendMessage(any(), any(), any())
-        verify(messageSender).sendMessage(channelType, channelId, message)
+        verify(sendMessageInterceptor).sendMessage(channelType, channelId, message)
     }
 }
